+++ src/components/editors/DimensionOrMeasureMultiSelectFieldPro/definition.ts
@@ -1,8 +1,8 @@
 import { Value } from '@embeddable.com/core';
 import { EmbeddedComponentMeta, Inputs, definePreview } from '@embeddable.com/react';
-import Component from './DimensionOrMeasureMultiSelectFieldPro';
 import { inputs } from '../../component.inputs.constants';
 import { previewData } from '../../preview.data.constants';
+import Component from './index';
 
 const meta = {
   name: 'DimensionMeasureMultiSelectFieldPro',
@@ -67,8 +67,6 @@ const preview = definePreview(Component, {
 const props = (inputs: Inputs<typeof meta>) => {
   return {
     ...inputs,
-    dimensionsAndMeasures: inputs.dimensionsAndMeasures ?? [],
-    selectedDimensionsAndMeasures: inputs.selectedDimensionsAndMeasures ?? [],
   };
 };
